requirements.txt
vyper==0.4.1
web3==6.11.0
python-dotenv==1.0.0
pytest==7.4.2
eth-account==0.9.0
